refactor: code style/format
