+++ src/main/java/fr/hiit/ipaidthat/configuration/IPTErrorDecoder.java
@@ -12,7 +12,7 @@ public class IPTErrorDecoder implements ErrorDecoder {
     public Exception decode(String methodKey, Response response) {
         HttpStatus status = HttpStatus.resolve(response.status());
         return status != null ?
-                new ResponseStatusException(status, response.reason()) :
-                new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Generic error");
+            new ResponseStatusException(status, response.reason()) :
+            new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Generic error");
     }
 }
