+++ src/main/java/fr/hiit/ipaidthat/proxy/model/Customer.java
@@ -2,12 +2,13 @@ package fr.hiit.ipaidthat.proxy.model;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import java.time.LocalDate;
-import javax.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.Size;
+import java.time.LocalDate;
 
 @Data
 @AllArgsConstructor
@@ -16,37 +17,36 @@ import lombok.NoArgsConstructor;
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
 public class Customer {
 
-  private Long id;
+    private Long id;
 
-  private LocalDate created;
+    private LocalDate created;
 
-  @Size(max = 64)
-  private String externalId;
+    @Size(max = 64)
+    private String externalId;
 
-  @Size(max = 64)
-  private String name;
+    @Size(max = 64)
+    private String name;
 
-  @Size(max = 64)
-  private String firstName;
+    @Size(max = 64)
+    private String firstName;
 
-  @Size(max = 64)
-  private String lastName;
+    @Size(max = 64)
+    private String lastName;
 
-  @Size(max = 30)
-  private String siren;
+    @Size(max = 30)
+    private String siren;
 
-  @Size(max = 30)
-  private String vat;
+    @Size(max = 30)
+    private String vat;
 
-  @Size(max = 500)
-  private String address;
+    @Size(max = 500)
+    private String address;
 
-  @Size(max = 64)
-  private String email;
+    @Size(max = 64)
+    private String email;
 
-  @Size(max = 32)
-  private String phone;
+    @Size(max = 32)
+    private String phone;
 
-  private String extra;
-
+    private String extra;
 }
