+++ src/main/java/fr/hiit/ipaidthat/proxy/model/InvoiceSimpleItem.java
@@ -2,12 +2,13 @@ package fr.hiit.ipaidthat.proxy.model;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
 
 @Data
 @AllArgsConstructor
@@ -16,26 +17,25 @@ import lombok.NoArgsConstructor;
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
 public class InvoiceSimpleItem {
 
-  private Long id;
+    private Long id;
 
-  @NotNull
-  private Long invoice;
+    @NotNull
+    private Long invoice;
 
-  @NotNull
-  @Size(min = 1, max = 64)
-  private String name;
+    @NotNull
+    @Size(min = 1, max = 64)
+    private String name;
 
-  private String additionalInfo;
+    private String additionalInfo;
 
-  @NotNull
-  private Double unitPrice;
+    @NotNull
+    private Double unitPrice;
 
-  private Double quantity;
+    private Double quantity;
 
-  private Double taxPercent;
+    private Double taxPercent;
 
-  private Double discountPercent;
+    private Double discountPercent;
 
-  private Integer position;
-
+    private Integer position;
 }
